[Fix] 필터 탠스택 쿼리 캐싱
