+++ src/pages/Filter/fetchFilter.js
@@ -1,0 +1,7 @@
+import pb from '@/api/pocketbase';
+
+export const fetchFilter = () => async () => {
+  const filters = await pb.collection('genres').getFullList();
+
+  return filters;
+};
