+++ src/pages/Filter/index.js
@@ -1,2 +1,3 @@
 export * from './Filter';
 export * from './loader';
+export * from './fetchFilter';
